pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "splatscout"
version = "0.1.0"
description = "Ubuntu-first CLI validation tool for 3D Gaussian Splatting datasets."
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
  {name = "Carlo"}
]
dependencies = [
  "pillow>=10.0",
  "numpy>=1.26",
  "typer>=0.12",
  "rich>=13.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]
optional = [
  "imagehash>=4.3",
]

[project.scripts]
splatscout = "splatscout.cli:main"

[tool.setuptools.packages.find]
include = ["splatscout*"]
